refactor the release command to use individual files per cmd
