+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/release/__init__.py
@@ -1,0 +1,24 @@
+# (C) Datadog, Inc. 2019
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+import click
+
+from ..console import CONTEXT_SETTINGS
+from .build import build
+from .changelog import changelog
+from .make import make
+from .show import show
+from .tag import tag
+from .testable import testable
+from .upload import upload
+
+ALL_COMMANDS = [build, changelog, make, show, tag, testable, upload]
+
+
+@click.group(context_settings=CONTEXT_SETTINGS, short_help='Manage the release of checks')
+def release():
+    pass
+
+
+for command in ALL_COMMANDS:
+    release.add_command(command)
